+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.2
 - Refractoring, add all streamwish model
+- Add streamwish upload and download extension
 
 ## 1.0.1
 - Add feature(s)